moved getVexString to Music model
