+++ app/javascript/controllers/score_controller.js
@@ -9,6 +9,7 @@
 
 import Vex from 'vexflow'
 import { Controller } from "stimulus"
+import { Music } from "../models/music"
 
 export default class extends Controller {
   static targets = [ "output" ]
@@ -17,13 +18,12 @@ export default class extends Controller {
   };
 
   connect() {
+    this.music = new Music(this.notesValue, 0)
     this.initConverters();
     this.currentSelection = null;
     const VF = Vex.Flow;
     this.vf = new Vex.Flow.Factory({renderer: {elementId: 'score'}})
     console.log("notes", this.notesValue)
-
-    this.noteEvents = JSON.parse(this.notesValue);
     console.log("lengths list", this.noteLengthList);
 
     this.draw();
@@ -52,21 +52,7 @@ export default class extends Controller {
   }
 
 
-  getVexString() {
-    const vexEvents = []
-    this.noteEvents.forEach((note, i) => {
-      if (Array.isArray(note[0])) {
-        if (note[0][0] == 'r') { // rest
-          vexEvents.push(`(${note[0][1]})/r${note[1]}`);
-        } else { //chord
-          vexEvents.push(`(${note[0].join(' ')})/${note[1]}`);
-        }
-      } else { // single note
-        vexEvents.push(`${note[0]}/${note[1]}`);
-      }
-    });
-    return vexEvents.join(', ')
-  }
+
 
   draw(event) {
     if (event) {
@@ -78,11 +64,8 @@ export default class extends Controller {
     const system = this.vf.System();
     console.log(this.counter);
 
-    const vexString = this.getVexString();
-
-    console.log("vexString", vexString);
     system.addStave({
-      voices: [score.voice(score.notes(vexString))]
+      voices: [score.voice(score.notes(this.music.getVexString()))]
     }).addClef('treble').addTimeSignature('4/4');
 
     console.log("redraw");
@@ -111,10 +94,10 @@ export default class extends Controller {
   keyDownOnNote(event) {
     let newMidiNum;
     console.log("keydown", event.code, event, event.metaKey);
-    console.log("before, note events:", this.noteEvents);
+    console.log("before, note events:", this.music.notes);
     let svgNote = event.currentTarget;
     let index = this.noteIndex(svgNote);
-    const midiNum = this.noteName2MidiNum[this.noteEvents[index][0]]
+    const midiNum = this.noteName2MidiNum[this.music.notes[index][0]]
     const refMidiNums = {
       'KeyC': 12,
       'KeyD': 14,
@@ -152,9 +135,9 @@ export default class extends Controller {
         break;
       case 'Digit4': // 8th note
         // break both list
-        console.log("Bef, 8th note", this.noteEvents)
-        this.noteEvents.splice(index,0,[['r', 'A4'], 8])
-        console.log("Aft, 8th note", this.noteEvents)
+        console.log("Bef, 8th note", this.music.notes)
+        this.music.notes.splice(index,0,[['r', 'A4'], 8])
+        console.log("Aft, 8th note", this.music.notes)
         // insert a new rest
         // change the note durations
         // update display
@@ -169,7 +152,7 @@ export default class extends Controller {
   }
 
   selectNextNote(event, index, svgNote) {
-    index = Math.min(index + 1, this.noteEvents.length-1)
+    index = Math.min(index + 1, this.music.notes.length-1)
     console.log("svgNote", svgNote);
     console.log("index", svgNote);
     return this.changeSelection(event, index, svgNote)
@@ -184,24 +167,24 @@ export default class extends Controller {
   }
 
   isRest(index) {
-    Array.isArray(this.noteEvents[index][0]) && Array.isArray(this.noteEvents[index][0])
+    Array.isArray(this.music.notes[index][0]) && Array.isArray(this.music.notes[index][0])
   }
 
 
   updateNote(event, index, accidental,newMidiNum) {
     // Note: works only for single notes. Doesn't handle chords
     console.log("updateNote -----------");
-    console.log("before, note events:", this.noteEvents[index][0]);
+    console.log("before, note events:", this.music.notes[index][0]);
     if (!this.isRest(index)) {
       if (accidental == '#') {
-        this.noteEvents[index][0] = this.midiNum2NoteNameSharp[newMidiNum];
+        this.music.notes[index][0] = this.midiNum2NoteNameSharp[newMidiNum];
       } else if (accidental == 'b' || accidental == 'n') {
-        this.noteEvents[index][0] = this.midiNum2NoteNameFlat[newMidiNum];
+        this.music.notes[index][0] = this.midiNum2NoteNameFlat[newMidiNum];
       } else {
         throw new Error(`Unknown accidental: ${accidental}. Accepted values are '#' and 'b'.`)
       }
     }
-    console.log("after, note events:", this.noteEvents);
+    console.log("after, note events:", this.music.notes);
     console.log("----------");
 
     this.draw(event);
@@ -245,7 +228,7 @@ export default class extends Controller {
 
   updateAttemptStringPlayback(event) {
     const toneController = document.querySelector("#tone-controller");
-    toneController.dataset.toneAttemptValue = JSON.stringify(this.noteEvents)
+    toneController.dataset.toneAttemptValue = JSON.stringify(this.music.notes)
     console.log("toneController dataset:", toneController.dataset);
   }
